+++ test/e2e/features/pageObjects/workAllocation/checkYourAnswersPage.js
@@ -1,0 +1,8 @@
+
+
+class CheckYourAnswersPage{
+
+
+}
+
+module.exports = new CheckYourAnswersPage();
