+++ test/e2e/features/step_definitions/workallocation/caseRolesAccess.steps.js
@@ -8,6 +8,11 @@ const caseRolesAndAccessPage = require("../../pageObjects/workAllocation/caseRol
 
 const ArrayUtil = require('../../../utils/ArrayUtil');
 
+const chooseWhoExclusionIsForPage = require("../../pageObjects/workAllocation/chooseWhoExclusionIsForPage");
+const choosePersonRolePage = require("../../pageObjects/workAllocation/choosePersonRolePage");
+const findThePersonPage = require("../../pageObjects/workAllocation/findPersonPage");
+const describeExclusionPage = require("../../pageObjects/workAllocation/describeExclusionPage");
+const checkYourAnswersPage = require("../../pageObjects/workAllocation/checkYourAnswersPage");
 
 defineSupportCode(function ({ And, But, Given, Then, When }) {
 
@@ -27,6 +32,25 @@ defineSupportCode(function ({ And, But, Given, Then, When }) {
 
     Then('I validate for {string} table column {string} value is {string} in case roles and access page', async function (roleType,ColumnHeader,columnValue) {
         expect(await caseRolesAndAccessPage.getTableColumnValueForAccessRoleType(roleType, ColumnHeader)).to.includes(columnValue);
+    });
+
+    Then('I see Add an exclusion work flow page {string} is displayed', async function(workFlowPage){
+        let workFlowPageObject = null;
+
+        if (workFlowPage === 'Choose who the exclusion is for'){
+            expect(await chooseWhoExclusionIsForPage.isDisplayed()).to.be.true;
+        } else if (workFlowPage === "Choose the person's role") {
+            expect(await choosePersonRolePage.isDisplayed()).to.be.true;
+        } else if (workFlowPage === "Find the person") {
+            expect(await findThePersonPage.isDisplayed()).to.be.true;
+        } else if (workFlowPage === "Describe the exclusion") {
+            expect(await describeExclusionPage.isDisplayed()).to.be.true;
+        } else if (workFlowPage === "Check your answers") {
+            expect(await checkYourAnswersPage.isDisplayed()).to.be.true;
+        }else{
+            throw new Error(`work flow page "${workFlowPage}" is not recognised or not implemented in test step definition.`);
+        }
+        
     });
 
 });
