+++ test/ngIntegration/tests/features/workallocation2/caseDetailsTabs.feature/WAAndRolesAccessTabs.feature
@@ -1,7 +1,7 @@
 @ng
-Feature: Tasks and Roles and access tabs visibility
+Feature: WA Release 2: Case details - "Tasks", "Roles and access" tabs visibility
 
-    Scenario Outline: WA tab <TabLabel> is displayed <isDisplayed> in case details for user <UserType> <UserIdentifier>
+    Scenario Outline: WA tab "<TabLabel>"" is displayed "<isDisplayed>"" in case details for user "<UserType>" "<UserIdentifier>"
         Given I set MOCK with user "<UserIdentifier>" and roles "<Roles>"
         Given I start MockApp
         Given I navigate to home page
